renaming
Renamed PlanetVis.js to AdjacencyMatrix.js
