+++ frontend/src/Pages/vis.js
@@ -1,7 +1,7 @@
 import ThreadArcs from './ThreadArcs';
 import '../Css/visual.css';
 import React from 'react';
-import AdjacencyMatrix from './PlanetVis';
+import AdjacencyMatrix from './AdjacencyMatrix';
 
 
 class Vis extends React.Component {
